refactor: exports input container
