+++ nuntius-pro-app/src/design-system/components/Inputs/Anatomy/InputContainer.js
@@ -1,0 +1,10 @@
+import { Container } from './InputAnatomy.styles';
+
+export const InputContainer = ({ error = false, children, endAdornment }) => {
+  return (
+    <Container aria-invalid={error}>
+      {children}
+      {endAdornment}
+    </Container>
+  );
+};
